server/user: fix OAuth account disconnection when users have multiple ones for the same platform
