+++ server/polar/user/oauth_service.py
@@ -1,8 +1,7 @@
-from uuid import UUID
-
 import structlog
-from sqlalchemy import delete, func, select
+from sqlalchemy import func, select
 
+from polar.exceptions import PolarError
 from polar.kit.services import ResourceServiceReader
 from polar.logging import Logger
 from polar.models import OAuthAccount, User
@@ -10,6 +9,25 @@ from polar.models.user import OAuthPlatform
 from polar.postgres import AsyncSession
 
 log: Logger = structlog.get_logger()
+
+
+class OAuthError(PolarError): ...
+
+
+class OAuthAccountNotFound(OAuthError):
+    def __init__(self, platform: OAuthPlatform) -> None:
+        self.platform = platform
+        message = f"No {platform} OAuth account found for this user."
+        super().__init__(message, 404)
+
+
+class CannotDisconnectLastAuthMethod(OAuthError):
+    def __init__(self) -> None:
+        message = (
+            "Cannot disconnect this OAuth account as it's your only authentication method. "
+            "Please verify your email or connect another OAuth provider before disconnecting."
+        )
+        super().__init__(message, 400)
 
 
 class OAuthAccountService(ResourceServiceReader[OAuthAccount]):
@@ -23,44 +41,40 @@ class OAuthAccountService(ResourceServiceReader[OAuthAccount]):
         result = await session.execute(stmt)
         return result.scalars().one_or_none()
 
-    async def get_by_platform_and_user_id(
-        self, session: AsyncSession, platform: OAuthPlatform, user_id: UUID
-    ) -> OAuthAccount | None:
-        stmt = select(OAuthAccount).where(
-            OAuthAccount.platform == platform,
-            OAuthAccount.user_id == user_id,
-        )
-        result = await session.execute(stmt)
-        return result.scalars().one_or_none()
-
-    async def can_disconnect_oauth_account(
-        self, session: AsyncSession, user: User, oauth_account_id: UUID
-    ) -> bool:
-        stmt = select(func.count(OAuthAccount.id)).where(
-            OAuthAccount.user_id == user.id,
-            OAuthAccount.id != oauth_account_id,
-        )
-        active_oauth_count = await session.scalar(stmt)
-
-        if active_oauth_count == 0 and not user.email_verified:
-            return False
-
-        return True
-
-    async def disconnect_oauth_account(
-        self,
-        session: AsyncSession,
-        user: User,
-        oauth_account_id: UUID,
-        platform: OAuthPlatform,
+    async def disconnect_platform(
+        self, session: AsyncSession, user: User, platform: OAuthPlatform
     ) -> None:
-        log.info(
-            "oauth_account.disconnect",
-            oauth_account_id=oauth_account_id,
-            platform=platform,
+        oauth_accounts_statement = select(OAuthAccount).where(
+            OAuthAccount.platform == platform,
+            OAuthAccount.user_id == user.id,
         )
-        stmt = delete(OAuthAccount).where(OAuthAccount.id == oauth_account_id)
-        await session.execute(stmt)
+        oauth_account_result = await session.execute(oauth_accounts_statement)
+        # Some users have a buggy state with multiple OAuth accounts for the same platform
+        oauth_accounts = oauth_account_result.scalars().all()
+
+        if len(oauth_accounts) == 0:
+            raise OAuthAccountNotFound(platform)
+
+        other_accounts_count_statement = select(func.count(OAuthAccount.id)).where(
+            OAuthAccount.user_id == user.id,
+            OAuthAccount.id.not_in([oa.id for oa in oauth_accounts]),
+        )
+        other_accounts_count_result = await session.execute(
+            other_accounts_count_statement
+        )
+        other_accounts_count = other_accounts_count_result.scalar_one()
+
+        if other_accounts_count == 0 and not user.email_verified:
+            raise CannotDisconnectLastAuthMethod()
+
+        for oauth_account in oauth_accounts:
+            await session.delete(oauth_account)
+            log.info(
+                "oauth_account.disconnect",
+                oauth_account_id=oauth_account.id,
+                platform=platform,
+            )
+
         await session.flush()
